+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `soyhuce/next-ide-helper` will be documented in this file
 ### Added
 
 - Attributes with CastsInboundAttribues cast have database defined type.
+- Return type for `sole` method
 
 ### Fixed